+++ airsync-mac/Constants/Constants.swift
@@ -9,6 +9,6 @@ import Foundation
 
 // Constants.swift
 enum Defaults {
-    static let serverPort: UInt16 = 6996
+    static let serverPort: UInt32 = 6996
 }
